+++ src/components/asset-list/RecyclerAssetList2/profile-header/ProfileNameRow.tsx
@@ -15,7 +15,7 @@ import { useAccountProfile, useDimensions } from '@/hooks';
 import { useNavigation } from '@/navigation';
 import { abbreviateEnsForDisplay } from '@/utils/abbreviations';
 import Routes from '@rainbow-me/routes';
-import { addressCopiedToastAtom } from '@/screens/WalletScreen/WalletScreen';
+import { addressCopiedToastAtom } from '@/screens/WalletScreen';
 import { FloatingEmojis } from '@/components/floating-emojis';
 import { haptics } from '@/utils';
 import { Space } from '@/design-system/docs/system/tokens.css';
