+++ src/components/cards/ReceiveAssetsCard.tsx
@@ -5,7 +5,7 @@ import { useAccountProfile, useClipboard } from '@/hooks';
 import { ButtonPressAnimation } from '../animations';
 import { FloatingEmojis } from '@/components/floating-emojis';
 import { useRecoilState } from 'recoil';
-import { addressCopiedToastAtom } from '@/screens/WalletScreen/WalletScreen';
+import { addressCopiedToastAtom } from '@/screens/WalletScreen';
 import { haptics } from '@/utils';
 import { useNavigation } from '@/navigation';
 import Routes from '@/navigation/routesNames';
